requirements.txt
streamlit>=1.32
pdfplumber>=0.11
pypdf>=4.0
pandas>=2.2
numpy>=1.26
faiss-cpu>=1.8
sentence-transformers>=2.6
plotly>=5.19
python-dateutil>=2.9
requests>=2.31
openai>=1.40
google-genai>=1.0
